# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TrackStar API",
    "description": "Personal Task & Habit Tracker API",
    "version": "1.0.0",
    "contact": {
      "name": "TrackStar Team",
      "email": "support@trackstar.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
...
    "/api/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current user profile",
        "description": "Retrieve the profile information of the currently authenticated user.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update current user profile",
        "description": "Update the profile information of the currently authenticated user.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Updated user information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "John Doe"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe"
                },
                "timezone": {
                  "type": "string",
                  "example": "America/New_York"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete/deactivate user account",
        "description": "Deactivate the current user account. This is a soft delete operation.",
        "parameters": [
          {
            "name": "x-user-id",
            "in": "header",
            "description": "User ID for authentication",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/register": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register new user (OAuth callback)",
        "description": "Create a new user account or login existing user via OAuth provider.",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User registration information from OAuth provider",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "providerId": {
                  "type": "string",
                  "example": "12345"
                },
                "provider": {
                  "type": "string",
                  "example": "github"
                },
                "name": {
                  "type": "string",
                  "example": "John Doe"
                },
                "email": {
                  "type": "string",
                  "example": "john@example.com"
                },
                "username": {
                  "type": "string",
                  "example": "johndoe"
                },
                "avatar": {
                  "type": "string",
                  "example": "https://avatar.url.com/john.jpg"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "User ID for authentication (temporary for testing)"
      }
    }
  }
}